+++ src/features/post/ui/PostEditDialog.tsx
@@ -1,9 +1,8 @@
-import { useState } from "react"
 import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Input, Textarea } from "../../../shared/ui"
+import { usePostStore } from "../model/store"
 
-export const PostEditDialog: React.FC = ({}) => {
-  const [showEditDialog, setShowEditDialog] = useState(false)
-
+export const PostEditDialog = ({ showEditDialog, setShowEditDialog, updatePost }) => {
+  const { selectedPost, setSelectedPost } = usePostStore()
   return (
     <Dialog open={showEditDialog} onOpenChange={setShowEditDialog}>
       <DialogContent>
@@ -14,13 +13,13 @@ export const PostEditDialog: React.FC = ({}) => {
           <Input
             placeholder="제목"
             value={selectedPost?.title || ""}
-            onChange={(e) => setSelectedPost({ ...selectedPost, title: e.target.value })}
+            onChange={(e) => selectedPost && setSelectedPost({ ...selectedPost, title: e.target.value })}
           />
           <Textarea
             rows={15}
             placeholder="내용"
             value={selectedPost?.body || ""}
-            onChange={(e) => setSelectedPost({ ...selectedPost, body: e.target.value })}
+            onChange={(e) => selectedPost && setSelectedPost({ ...selectedPost, body: e.target.value })}
           />
           <Button onClick={updatePost}>게시물 업데이트</Button>
         </div>
