+++ src/shared/lib/highlightText.tsx
@@ -1,0 +1,18 @@
+import React from "react"
+
+export const highlightText = (text: string | undefined, highlight: string): React.JSX.Element | null => {
+  if (!text) return null
+
+  if (!highlight.trim()) {
+    return <span>{text}</span>
+  }
+
+  const regex = new RegExp(`(${highlight})`, "gi")
+  const parts = text.split(regex)
+
+  return (
+    <span>
+      {parts.map((part, i) => (regex.test(part) ? <mark key={i}>{part}</mark> : <span key={i}>{part}</span>))}
+    </span>
+  )
+}
